Change monitring methods

+++ Epam.Task05/Epam.Task05.BackupSystem/FileMonitor.cs
@@ -10,28 +10,34 @@ namespace Epam.Task05.BackupSystem
     {
         private readonly FileSystemWatcher fileSystemWatcher;
         private readonly string currentFolder;
+        private readonly string archivFolder;
         private readonly string backupFolder;
         private readonly string logFile;
         private readonly string filter;
         private StreamWriter log = null;
-        private int fileQuantity = 0;
+        private long id;
 
         public FileMonitor(string currentFolder, string backupFolder, string filter, bool monitorOn)
         {
+            
             this.currentFolder = currentFolder;
             this.backupFolder = backupFolder;
+            this.archivFolder = Path.Combine(this.backupFolder, "archiv");
             this.filter = filter;
             this.logFile = Path.Combine(backupFolder, "log.txt");
             if (!Directory.Exists(backupFolder) || !File.Exists(this.logFile))
             {
+                this.id = 0;
                 this.FirstStart();
             }
             else
             {
-                this.fileQuantity = Directory.GetFiles(this.backupFolder, "Backup*" + filter, SearchOption.AllDirectories).Length;
+                this.id = Directory.GetFiles(this.backupFolder, "Backup*" + filter, SearchOption.AllDirectories).Length + 1;
             }
 
             this.fileSystemWatcher = new FileSystemWatcher(currentFolder);
+            this.fileSystemWatcher.NotifyFilter = NotifyFilters.LastWrite | NotifyFilters.FileName | NotifyFilters.Attributes | NotifyFilters.CreationTime;
+            this.fileSystemWatcher.InternalBufferSize = 1000;
             this.fileSystemWatcher.IncludeSubdirectories = true;
             this.fileSystemWatcher.Changed += this.FileWatcherOnChanged;
             this.fileSystemWatcher.Created += this.FileWatcherOnCreated;
@@ -40,61 +46,57 @@ namespace Epam.Task05.BackupSystem
             this.fileSystemWatcher.EnableRaisingEvents = monitorOn;
         }
 
-        public void Recovery(DateTime recDateTime)
-        {
-            if (this.fileSystemWatcher.EnableRaisingEvents == true)
-            {
-                Console.WriteLine("Wile watcher in work you cann`t have recovery!");
-                return;
-            }
+        //public void Recovery(DateTime recDateTime)
+        //{
+        //    if (this.fileSystemWatcher.EnableRaisingEvents == true)
+        //    {
+        //        Console.WriteLine("Wile watcher in work you cann`t have recovery!");
+        //        return;
+        //    }
 
-            StreamReader log = null;
-            DateTime logData;
-            string action;
-            string[] allDirectoryFiles = Directory.GetFiles(this.currentFolder, "*" + this.filter, SearchOption.AllDirectories);
-            foreach (var file in allDirectoryFiles)
-            {
-                File.Delete(file);
-            }
+        //    StreamReader log = null;
+        //    DateTime logData;
+        //    string action;
+        //    string[] allDirectoryFiles = Directory.GetFiles(this.currentFolder, "*" + this.filter, SearchOption.AllDirectories);
+        //    foreach (var file in allDirectoryFiles)
+        //    {
+        //        File.Delete(file);
+        //    }
 
-            using (log = new StreamReader(this.logFile))
-            {
-                logData = DateTime.Parse(log.ReadLine());
-                if (recDateTime < logData)
-                {
-                    recDateTime = logData;
-                }
+        //    using (log = new StreamReader(this.logFile))
+        //    {
+        //        logData = DateTime.Parse(log.ReadLine());
+        //        if (recDateTime < logData)
+        //        {
+        //            recDateTime = logData;
+        //        }
 
-                while (logData <= recDateTime)
-                {
-                    action = log.ReadLine();
-                    this.FileRecovery(log, action);
-                    if (log.Peek() < 0)
-                    {
-                        break;
-                    }
+        //        while (logData <= recDateTime)
+        //        {
+        //            action = log.ReadLine();
+        //            this.FileRecovery(log, action);
+        //            if (log.Peek() < 0)
+        //            {
+        //                break;
+        //            }
 
-                    logData = DateTime.Parse(log.ReadLine());
-                }
+        //            logData = DateTime.Parse(log.ReadLine());
+        //        }
 
-                Console.WriteLine("Recovery finished");
-            }
-        }
+        //        Console.WriteLine("Recovery finished");
+        //    }
+        //}
 
-        private void SaveFile(string fileEvent, params string[] storedFilesName)
+        private void SaveFile(string fileEvent, params string[] filesName)
         {
             using (this.log = new StreamWriter(this.logFile, true))
             {
-                string data = DateTime.Now.ToString();
-                foreach (var fileName in storedFilesName)
+                foreach (var file in filesName)
                 {
-                    string backupFileName = Path.Combine(this.backupFolder, "Backup" + this.fileQuantity.ToString() + ".txt");
-                    File.Copy(fileName, Path.Combine(this.backupFolder, backupFileName));
-                    this.log.WriteLine(data);
-                    this.log.WriteLine(fileEvent);
-                    this.log.WriteLine(fileName);
-                    this.log.WriteLine(backupFileName);
-                    this.fileQuantity++;
+                    string backupFileName = Path.Combine(this.archivFolder, "Backup" + this.id.ToString() + ".txt");
+                    File.Copy(file, Path.Combine(this.archivFolder, backupFileName));
+                    this.log.WriteLine(id + "|||" + DateTime.Now.ToString() + "|||" + fileEvent + "|||" + file + "|||" + backupFileName + "|!|");
+                    this.id++;
                 }
             }
 
@@ -103,46 +105,35 @@ namespace Epam.Task05.BackupSystem
 
         private void FileWatcherOnRenamed(object sender, RenamedEventArgs systemEvent)
         {
-            if (!File.Exists(systemEvent.FullPath))
+            if (!systemEvent.FullPath.Contains(this.currentFolder) || !validationExtension(systemEvent))
             {
+                FileWatcherOnDeleted(sender, systemEvent);
                 return;
             }
 
-            var fileInfo = new FileInfo(systemEvent.OldFullPath);
-            if (fileInfo.Extension != this.filter)
-            {
-                return;
-            }
-
-            Console.WriteLine("File: {0} renamed to {1}", systemEvent.OldFullPath, systemEvent.FullPath);
-            string newFileName = systemEvent.FullPath;
-            string oldFileName = systemEvent.OldFullPath;
+            Console.WriteLine($"File: {systemEvent.OldFullPath} renamed to {systemEvent.FullPath}");
             using (this.log = new StreamWriter(this.logFile, true))
             {
-                this.log.WriteLine(DateTime.Now.ToString());
-                this.log.WriteLine("Renamed");
-                this.log.WriteLine(newFileName);
-                this.log.WriteLine(oldFileName);
+                this.log.WriteLine(id + "|||" + DateTime.Now.ToString() + "|||" + "Renamed" + "|||" + systemEvent.FullPath + "|||" + systemEvent.OldFullPath + "|!|");
             }
 
             this.log.Close();
         }
 
-        private void FileWatcherOnDeleted(object sender, FileSystemEventArgs systemEvent)
+        private bool validationExtension(FileSystemEventArgs systemEvent)
         {
             var fileInfo = new FileInfo(systemEvent.FullPath);
-            if (fileInfo.Extension != this.filter)
-            {
-                return;
-            }
+            string fileExtension = fileInfo.Extension.Substring(fileInfo.Extension.LastIndexOf('.') + 1);
+            string filterExtension = this.filter.Substring(this.filter.LastIndexOf('.') + 1);
+            return fileExtension == filterExtension;
+        }
 
+        private void FileWatcherOnDeleted(object sender, FileSystemEventArgs systemEvent)
+        {
             Console.WriteLine("File: {0} deleted", systemEvent.FullPath);
-            string deletedFileName = systemEvent.FullPath;
             using (this.log = new StreamWriter(this.logFile, true))
             {
-                this.log.WriteLine(DateTime.Now.ToString());
-                this.log.WriteLine("Deleted");
-                this.log.WriteLine(deletedFileName);
+                this.log.WriteLine(id + "|||" + DateTime.Now.ToString() + "|||" + "Deleted" + "|||" + systemEvent.FullPath + "|!|");
             }
 
             this.log.Close();
@@ -150,46 +141,36 @@ namespace Epam.Task05.BackupSystem
 
         private void FileWatcherOnCreated(object sender, FileSystemEventArgs systemEvent)
         {
-            if (!File.Exists(systemEvent.FullPath))
-            {
-                return;
-            }
+            //if (!File.Exists(systemEvent.FullPath))
+            //{
+            //    return;
+            //}
 
-            var fileInfo = new FileInfo(systemEvent.FullPath);
-            if (fileInfo.Extension != this.filter)
+            if (!validationExtension(systemEvent))
             {
                 return;
             }
 
             Console.WriteLine("File: {0} created", systemEvent.FullPath);
-            string createdFileName = systemEvent.FullPath;
-            this.SaveFile("Created", createdFileName);
+            this.SaveFile("Created", systemEvent.FullPath);
         }
 
         private void FileWatcherOnChanged(object sender, FileSystemEventArgs systemEvent)
         {
-            try
+            if (!systemEvent.FullPath.Contains(this.currentFolder) || !validationExtension(systemEvent))
             {
-                this.fileSystemWatcher.EnableRaisingEvents = false;
-                if (!File.Exists(systemEvent.FullPath))
-                {
-                    return;
-                }
-
-                var fileInfo = new FileInfo(systemEvent.FullPath);
-                if (fileInfo.Extension != this.filter)
-                {
-                    return;
-                }
-
-                Console.WriteLine("File: {0} changed", systemEvent.FullPath);
-                string storedFileName = systemEvent.FullPath;
-                this.SaveFile("Changed", storedFileName);
+                FileWatcherOnDeleted(sender, systemEvent);
+                return;
             }
-            finally
-            {
-                this.fileSystemWatcher.EnableRaisingEvents = true;
-            }
+
+            Console.WriteLine("File: {0} changed", systemEvent.FullPath);
+            //using (this.log = new StreamWriter(this.logFile, true))
+            //{
+                //this.log.WriteLine(id + "|||" + DateTime.Now.ToString() + "|||" + "Changed" + "|||" + systemEvent.FullPath + "|!|");
+                this.SaveFile("Changed", systemEvent.FullPath);
+            //}
+
+            //this.log.Close();
         }
 
         private void MonitorOn()
@@ -249,10 +230,13 @@ namespace Epam.Task05.BackupSystem
             }
         }
 
-        private void FullBackup()
+        private void CreateArchiv()
         {
             string[] allDirectoryFiles = Directory.GetFiles(this.currentFolder, "*" + this.filter, SearchOption.AllDirectories);
-            this.SaveFile("Created", allDirectoryFiles);
+            if (allDirectoryFiles != null)
+            {
+                this.SaveFile("Created", allDirectoryFiles);
+            }
         }
 
         private void FirstStart()
@@ -261,35 +245,53 @@ namespace Epam.Task05.BackupSystem
             {
                 Directory.CreateDirectory(this.backupFolder);
             }
-            catch (ArgumentException argEx)
+            catch (ArgumentException ex)
             {
-                Console.WriteLine(argEx.Message);
+                Console.WriteLine(ex.Message);
+            }
+            catch (DirectoryNotFoundException ex)
+            {
+                Console.WriteLine(ex.Message);
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.Message);
             }
 
-            string[] allDirectoryFiles = Directory.GetFiles(this.backupFolder);
-            foreach (var file in allDirectoryFiles)
+            if (this.backupFolder == this.currentFolder)
             {
-                File.Delete(file);
+                throw new Exception("You cann`t watching backup folder!");
+            }
+
+            if (Directory.Exists(this.archivFolder))
+            {
+                if (Directory.GetFiles(this.archivFolder, "Backup*.txt") != null)
+                {
+                    foreach (var file in Directory.GetFiles(this.archivFolder, "Backup*.txt"))
+                    {
+                        File.Delete(file);
+                    }
+                }
+            }
+            else
+            {
+                Directory.CreateDirectory(this.archivFolder);
             }
 
             try
             {
                 File.Create(this.logFile).Close();
             }
-            catch (ArgumentException argEx)
+            catch (ArgumentException ex)
             {
-                Console.WriteLine(argEx.Message);
+                Console.WriteLine(ex.Message);
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.Message);
             }
 
-            this.FullBackup();
+            this.CreateArchiv();
         }
     }
 }
